cv-people-counter/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
paho-mqtt>=2.0.0
numpy>=1.24.0
